reuse the long-term key for the AAD part of the tag
This gives the implementation the option of pre-processing the AAD, if
all or part of it is constant between messages. To domain-separate the
AAD tag from the stream, we use a large initial value of the block
counter for that invocation. This reduces the maximum supported message
and AAD lengths from 2^64-1 to 2^63.

I don't love this approach, because doing domain separation by chopping
the XOF space in half feels like a hack. Doing another key derivation
step would be cleaner, but I don't want to pay the cost of an extra
compression if I don't have to.