avoid deriving any subkeys
This change reduces the max message and AAD lengths from 2^63 to 2^62.

This is doubling down on the "seek as domain separation" strategy, which
again I don't love, but it avoids spending any extra blocks on key
derivation. A <=48 byte message with no AAD now requires only two
compressions. On the decryption side, those two compressions can be done
in parallel.